rename: 람다 함수 이름 변경
