+++ backend/storage/src/tests/getUserFiles.mjs
@@ -1,7 +1,7 @@
-// getUserUsage.test.mjs
+// getUserFiles.mjs
 import AWSMock from 'aws-sdk-mock';
 import AWS from 'aws-sdk';
-import { handler } from '../handlers/getUserUsage.mjs';
+import { handler } from '../handlers/getUserFiles.mjs';
 
 // AWS SDK Mock 설정
 AWSMock.setSDKInstance(AWS);
